requirements.txt
google-cloud-bigquery==3.25.0
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.0